@@ -1,24 +1,25 @@
 {
     "name": "yma-config-plugin-typescript",
-    "version": "1.0.5",
+    "version": "1.0.6",
     "description": "",
     "main": "./dist/index.js",
     "scripts": {
         "init": "rm -rf package-lock.json && npm install",
         "start": "npm run build -- -w",
         "build": "npm run clean && tsc -p src",
-        "clean": "rm -fr dist"
+        "clean": "rm -fr dist",
+        "link": "npm link"
     },
     "files": ["bin", "dist", "public"],
     "dependencies": {
         "@babel/core": "7.17.8",
         "babel-loader": "8.2.4",
         "ts-loader": "9.2.8",
-        "typescript": "4.2.3"
+        "typescript": "5.1.6"
     },
     "devDependencies": {
         "webpack": "5.70.0",
-        "yma-config-plugin": "^1.0.5"
+        "yma-config-plugin": "^1.0.6"
     },
     "keywords": [],
     "author": "",
