@@ -68,7 +68,7 @@ module.exports = function ({types: t}) {
 
                     // TODO 输出文件路径
                     console.warn(
-                        `[babel-webpack-chunkname-plugin]: 通过 [${arg.value}] 生成 WebpackChunkName [${webpackChunkName}]，建议手动调整更好的应用 HTTP Cache`,
+                        `[babel-webpack-chunkname-plugin]: 通过 [${arg.value}] 生成 WebpackChunkName [${webpackChunkName}]，建议手动调整更好的应用 HTTP Cache`
                     );
                 }
             },
