@@ -42,7 +42,8 @@ module.exports = class extends Chainable {
     when(condition, whenTruthy = Function.prototype, whenFalsy = Function.prototype) {
         if (condition) {
             whenTruthy(this);
-        } else {
+        }
+        else {
             whenFalsy(this);
         }
 
