@@ -38,5 +38,5 @@ module.exports = Orderable(
 
             return config;
         }
-    },
+    }
 );
