@@ -1,4 +1,3 @@
 export function isDOMError(value) {
     return Object.prototype.toString.call(value) === '[object DOMError]';
-  }
-  
+}
