@@ -1,3 +1,3 @@
 export function isFalsy(v) {
     return v ? false : true;
-};
+}
