@@ -1,3 +1,3 @@
 export function isFalse(v) {
     return v === false;
-};
+}
