@@ -1,3 +1,3 @@
 export function isFormData(v) {
     return typeof FormData !== 'undefined' && v instanceof FormData;
-};
+}
