@@ -1,3 +1,3 @@
 export function isIOS(ua) {
     return ua && /iphone|ipad|ipod|ios/.test(ua);
-};
+}
