@@ -2,8 +2,5 @@ import {isObjectLike} from './isObjectLike';
 import {toString} from './constant';
 
 export function isNumber(v) {
-    return (
-        typeof v == 'number' ||
-        (isObjectLike(v) && toString.call(v) == '[object Number]')
-    );
+    return typeof v == 'number' || (isObjectLike(v) && toString.call(v) == '[object Number]');
 }
