@@ -2,4 +2,4 @@ export function isPrimitive(v) {
     const type = typeof v;
 
     return ['string', 'number', 'symbol', 'boolean'].includes(type);
-};
+}
