@@ -5,10 +5,6 @@ export function isString(v) {
     const type = typeof v;
 
     return (
-        type === 'string' ||
-        (type === 'object' &&
-            v != null &&
-            !isArray(v) &&
-            toString.call(v) === '[object String]')
+        type === 'string' || (type === 'object' && v != null && !isArray(v) && toString.call(v) === '[object String]')
     );
 }
