@@ -2,7 +2,6 @@ export function isSupportPointerEvents() {
     return !!(
         window.navigator['pointerEnabled'] ||
         window.PointerEvent ||
-        ('maxTouchPoints' in window.navigator &&
-            window.navigator.maxTouchPoints > 0)
+        ('maxTouchPoints' in window.navigator && window.navigator.maxTouchPoints > 0)
     );
 }
