@@ -5,8 +5,7 @@ export function isSupportTouch() {
             return !!(
                 window.navigator.maxTouchPoints > 0 ||
                 'ontouchstart' in window ||
-                (window['DocumentTouch'] &&
-                    document instanceof window['DocumentTouch'])
+                (window['DocumentTouch'] && document instanceof window['DocumentTouch'])
             );
         })()
     );
