@@ -1,5 +1,3 @@
 export function isURLSearchParams(v) {
-    return (
-        typeof URLSearchParams !== 'undefined' && v instanceof URLSearchParams
-    );
+    return typeof URLSearchParams !== 'undefined' && v instanceof URLSearchParams;
 }
